feat(consensus): CONS-501b — migrate codec, heartbeat, validator_protocol into lib-consensus-net
Completes the CONS-501 module split. The CONS-201 Scope B type relocation
(#2425) unblocked these — the moved modules transitively reference
ConsensusProposal/ConsensusVote/ConsensusProof, all of which now live
in lib-consensus-core.

## Moves (3,125 LOC across 3 files)

- `lib-consensus/src/network/codec.rs`         (813 LOC) → `lib-consensus-net/src/codec/mod.rs`
- `lib-consensus/src/network/heartbeat.rs`     (849 LOC) → `lib-consensus-net/src/heartbeat/mod.rs`
- `lib-consensus/src/validators/validator_protocol.rs` (1463 LOC) → `lib-consensus-net/src/validator_protocol/mod.rs`

## Re-exports keep callers compiling

`lib-consensus/src/network/mod.rs` and
`lib-consensus/src/validators/mod.rs` declare wrapper modules:

```rust
pub mod codec { pub use lib_consensus_net::codec::*; }
pub mod heartbeat { pub use lib_consensus_net::heartbeat::*; }
pub mod validator_protocol { pub use lib_consensus_net::validator_protocol::*; }
```

so every existing import (`lib_consensus::network::HeartbeatTracker`,
`lib_consensus::validators::validator_protocol::ValidatorMessage`, etc.)
keeps working unchanged. The engine, integration tests, and external
crates (zhtp, lib-blockchain, lib-network) all compile without
edits.

## Internal path rewrites in moved files

- `crate::types::ConsensusStep` → `lib_types::consensus::ConsensusStep`
- `crate::types::{ConsensusProposal, ConsensusVote, VoteType}`
  → `lib_consensus_core::types::{ConsensusProposal, ConsensusVote}` +
  `lib_types::consensus::VoteType`
- `crate::validators::ValidatorDiscoveryProtocol` →
  `crate::discovery::ValidatorDiscoveryProtocol`
- `crate::validators::validator_protocol::*` (self-references in
  cross-file callers) → `crate::validator_protocol::*`
- `crate::types::ValidatorMessage` (self-reference in
  validator_protocol's own tests) → `ValidatorMessage`
- `super::ValidatorAnnouncement` / `super::ValidatorStatus` (test
  references via parent `validators::*`) →
  `crate::discovery::ValidatorAnnouncement` /
  `crate::discovery::ValidatorStatus`

## Sign helpers visibility

`sign_propose_envelope`, `sign_vote_envelope`,
`sign_heartbeat_envelope` were `pub(crate)` in the old location —
their only callers were in the same crate. After the move,
`lib-consensus`'s `state_machine.rs` calls them across crate
boundaries, so they're now `pub`.

## Test sites that needed adjusting

- `codec/mod.rs` test: `crate::engines::consensus_engine::
  CONSENSUS_PROTOCOL_VERSION` was an internal lib-consensus path. The
  test now hard-codes the value `2` with a comment pointing at the
  authoritative constant in `lib-consensus`. If the engine bumps the
  version, both sides need to agree — the assertion will catch the
  drift.
- `lib-consensus-net/Cargo.toml` adds `rand = "0.8"` to
  `dev-dependencies` (the codec round-trip fuzz test uses it).

## Validation

- `cargo build --workspace` — clean (38.95 s warm).
- `cargo test -p lib-consensus-net --lib` — **53/53** pass (was 6;
  47 new from the moved modules).
- `cargo test -p lib-consensus --lib` — **134/134** pass (was 181;
  47 tests moved with the files).
- `cargo test -p lib-consensus-core --lib` — **41/41**.
- 134 + 53 = 187 — same total as pre-CONS-501b. No tests lost.

## Acceptance against #2351

- ✅ All four modules (discovery + codec + heartbeat +
  validator_protocol) compile in `lib-consensus-net`.
- ✅ External crates: zero deep-imports of
  `lib_consensus::validators::validator_discovery` /
  `lib_consensus::validators::validator_protocol` /
  `lib_consensus::network::*`. (Internal lib-consensus tests keep
  using their own re-exports — by design, no churn needed.)

Closes #2351
