+++ lib-consensus-net/Cargo.toml
@@ -31,3 +31,4 @@ bincode            = "1.3"
 
 [dev-dependencies]
 tokio              = { version = "1.0", features = ["full"] }
+rand               = "0.8"
