+++ lib-consensus/src/network/mod.rs
@@ -9,8 +9,14 @@
 //! It ensures deterministic serialization/deserialization of consensus messages
 //! while remaining opaque to message semantics.
 
-pub mod codec;
-pub mod heartbeat;
+// CONS-501b: codec + heartbeat migrated to lib-consensus-net.
+// Re-export modules so existing internal callers compile unchanged.
+pub mod codec {
+    pub use lib_consensus_net::codec::*;
+}
+pub mod heartbeat {
+    pub use lib_consensus_net::heartbeat::*;
+}
 pub mod liveness_monitor;
 
 pub use codec::{BincodeConsensusCodec, CodecError, ConsensusMessageCodec};
